image_types_qcom: generate flashable images for spinor
Spinor firmware is needed on iq-x7181 based devices, and recently
included as part of hamoa sprint one release, deploy them to generate
the flat build for spinor flash. Meanwhile, copy the programer outside
of spinor folder to be part of NVME flat build.

+++ classes-recipe/image_types_qcom.bbclass
@@ -19,6 +19,7 @@ QCOM_DTB_FILE ?= "dtb.bin"
 
 QCOM_BOOT_FILES_SUBDIR ?= ""
 QCOM_PARTITION_FILES_SUBDIR ??= "${QCOM_BOOT_FILES_SUBDIR}"
+QCOM_PARTITION_FILES_SUBDIR_SPINOR ??= ""
 
 QCOM_PARTITION_CONF ?= "qcom-partition-conf"
 
@@ -87,7 +88,9 @@ create_qcomflash_pkg() {
         install -m 0644 ${rawpg} .
     done
 
-    if [ -n "${QCOM_CDT_FILE}" ]; then
+    # check CDT file existence and install as per need, for iq-x7181, CDT file will be part of spinor build, which will
+    # be placed under spinor subfolder.
+    if [ -n "${QCOM_CDT_FILE}" ] && [ -e "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/${QCOM_CDT_FILE}.bin" ]; then
         install -m 0644 ${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/${QCOM_CDT_FILE}.bin cdt.bin
     fi
 
@@ -126,10 +129,42 @@ create_qcomflash_pkg() {
         install -m 0644 "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/${xbl_config}" xbl_config.elf
     fi
 
+    # programmer
+    if [ -e "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/xbl_s_devprg_ns.melf" ]; then
+        install -m 0644 "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/xbl_s_devprg_ns.melf" .
+    fi
+
     # sail nor firmware
     if [ -d "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/sail_nor" ]; then
         install -d sail_nor
         find "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/sail_nor" -maxdepth 1 -type f -exec install -m 0644 {} sail_nor \;
+    fi
+
+    # spinor firmware, partition bins, CDT etc.
+    if [ -d "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/spinor" ]; then
+        install -d spinor
+        find "${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/spinor" -maxdepth 1 -type f -exec install -m 0644 {} spinor \;
+
+        # partition bins for spinor
+        for pbin in `find ${DEPLOY_DIR_IMAGE}/${QCOM_PARTITION_FILES_SUBDIR_SPINOR} -maxdepth 1 -type f -name 'gpt_main*.bin' \
+                    -o -name 'gpt_backup*.bin' -o -name 'patch*.xml'`; do
+            install -m 0644 ${pbin} spinor
+        done
+
+        # skip BLANK_GPT and WIPE_PARTITIONS for rawprogram xml files
+        for rawpg in `find ${DEPLOY_DIR_IMAGE}/${QCOM_PARTITION_FILES_SUBDIR_SPINOR} -maxdepth 1 -type f -name 'rawprogram*.xml' \
+                    ! -name 'rawprogram*_BLANK_GPT.xml' ! -name 'rawprogram*_WIPE_PARTITIONS.xml'`; do
+            install -m 0644 ${rawpg} spinor
+        done
+
+        for zeros in `find ${DEPLOY_DIR_IMAGE}/${QCOM_PARTITION_FILES_SUBDIR_SPINOR} -maxdepth 1 -type f -name 'zeros_*.bin'`; do
+            install -m 0644 ${zeros} spinor
+        done
+
+        # rename CDT file to match the filename in partition conf
+        if [ -n "${QCOM_CDT_FILE}" ]; then
+            install -m 0644 ${DEPLOY_DIR_IMAGE}/${QCOM_BOOT_FILES_SUBDIR}/spinor/${QCOM_CDT_FILE}.bin spinor/cdt.bin
+        fi
     fi
 
     # Create symlink to ${QCOMFLASH_DIR} dir
